+++ marketing/src/App.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { HomePage } from './pages/HomePage';
-
-export function App() {
-  return <HomePage />;
-}